Add more backbone
Add base_agent, data, container to common and create physicore exe

+++ common/include/types.h
@@ -5,7 +5,6 @@
 namespace physicore {
 
 using real_t = double;
-
-using index_t = std::int64_t;
+using index_t = std::size_t;
 
 } // namespace physicore
